+++ src/Providers/Tenants/DatabaseDriverProvider.php
@@ -8,7 +8,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @see https://laravel-tenancy.com
+ * @see https://tenancy.dev
  * @see https://github.com/hyn/multi-tenant
  */
 
@@ -32,7 +32,7 @@ class DatabaseDriverProvider extends ServiceProvider
         $this->app->singleton(DatabaseDriverFactory::class);
     }
 
-    private function drivers()
+    protected function drivers()
     {
         $isPgsqlSchema = config('tenancy.db.tenant-division-mode') === Connection::DIVISION_MODE_SEPARATE_SCHEMA;