+++ src/Providers/Tenants/QueueProvider.php
@@ -8,7 +8,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @see https://laravel-tenancy.com
+ * @see https://tenancy.dev
  * @see https://github.com/hyn/multi-tenant
  */
 
@@ -17,51 +17,45 @@ namespace Hyn\Tenancy\Providers\Tenants;
 use Hyn\Tenancy\Contracts\Repositories\WebsiteRepository;
 use Hyn\Tenancy\Environment;
 use Illuminate\Queue\Events\JobProcessing;
-use Illuminate\Support\ServiceProvider;
 use Illuminate\Queue\QueueManager;
+use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Arr;
+use Hyn\Tenancy\Queue\DispatcherMiddleware;
+use Illuminate\Contracts\Bus\Dispatcher;
 
 class QueueProvider extends ServiceProvider
 {
-    public function register()
+    public function boot()
     {
-        $this->app->extend('queue', function (QueueManager $queue) {
-            $queue->createPayloadUsing(function (string $connection, string $queue = null, array $payload = []) {
-                if (isset($payload['website_id'])) {
-                    return [];
-                }
-
-                /** @var Environment $environment */
-                $environment = resolve(Environment::class);
-                $tenant = $environment->tenant();
-
-                return $tenant ? [
-                    'website_id' => $tenant->id,
-                ] : [];
-            });
-
-            $queue->before(function (JobProcessing $event) {
-                /** @var array $payload */
-                $payload = $event->job->payload();
-                if ($command = Arr::get($payload, 'data.command')) {
-                    $command = unserialize($command);
-                }
-
-                $key = $command->website_id ?? $payload['website_id'] ?? null;
-
-                if ($key) {
+        $this->app->booted(function () {
+            $this->app->extend('queue', function (QueueManager $queue) {
+                $queue->createPayloadUsing(function (string $connection, string $queue = null, array $payload = []) {
                     /** @var Environment $environment */
                     $environment = resolve(Environment::class);
-                    /** @var WebsiteRepository $repository */
-                    $repository = resolve(WebsiteRepository::class);
 
-                    $tenant = $repository->findById($key);
+                    /** @var mixed|null $website_id */
+                    $website_id = Arr::get($payload, 'data.command')->website_id ?? optional($environment->tenant())->getKey();
 
-                    $environment->tenant($tenant);
-                }
+                    return ['website_id' => $website_id];
+                });
+
+                return $queue;
             });
-
-            return $queue;
         });
+
+        $this->app['events']->listen(JobProcessing::class, function ($event) {
+            if ($key = Arr::get($event->job->payload(), 'website_id')) {
+                /** @var Environment $environment */
+                $environment = resolve(Environment::class);
+                /** @var WebsiteRepository $repository */
+                $repository = resolve(WebsiteRepository::class);
+
+                $tenant = $repository->findById($key);
+
+                $environment->tenant($tenant);
+            }
+        });
+
+        $this->app->make(Dispatcher::class)->pipeThrough([DispatcherMiddleware::class]);
     }
 }